// studform-master/Backend/studform-api/src/modules/students/dto/student.dto.ts
export class StudentDto {
    studentID: number;
    FirstName: string;
    LastName: string;
    MiddleName: string;
    FullName: string;
    Dob: string; // Assuming you'll handle dates as strings for simplicity
    Age: number;
    Email: string;
    PhoneNumber: string;
    Address: string;
    Gender: string;
    Course: string;
  }